Moved ArgsEncode and ArgsDecode out of OptionAll.. Will be deprecated
